Added spread guns, distance degradation on damage. Spread guns and penetration guns are now both integrated with standard guns, for flexibility purposes.
git-svn-id: https://legesmotus.svn.sourceforge.net/svnroot/legesmotus/legesmotus/trunk@1059 a5a5d5d7-22d1-44d1-b37a-67d3d99fcfd7

+++ common/Weapon.hpp
@@ -57,7 +57,7 @@ namespace LM {
 		// TODO: Do we want these to be in the common Weapon, or just the client one?
 		// When the current player (player) fires this weapon:
 		//  direction is in radians
-		virtual void		fire(b2World* physics, Player& player, Point start, double direction) = 0;
+		virtual void		fire(b2World* physics, Player& player, Point start, float direction) = 0;
 		// When another player (player) fires this weapon:
 		//virtual void		discharged(Player& player, StringTokenizer& data) = 0;
 		// When the current player (player) is hit by this weapon:
